return values in { 1 : <value> } format still to do: all applications

+++ api/oc_core_res.c
@@ -234,10 +234,10 @@ oc_core_get_num_devices(void)
 }
 
 int
-oc_core_set_device_fwv(int device_index, int major, int minor, int minor2)
+oc_core_set_device_fwv(size_t device_index, int major, int minor, int minor2)
 {
   if (device_index >= (int)oc_number_of_devices()) {
-    PRINT("device_index %d to large\n", device_index);
+    PRINT("device_index %d to large\n", (int)device_index);
     return -1;
   }
   oc_device_info[device_index].fwv.major = major;
@@ -247,10 +247,10 @@ oc_core_set_device_fwv(int device_index, int major, int minor, int minor2)
 }
 
 int
-oc_core_set_device_hwv(int device_index, int major, int minor, int minor2)
+oc_core_set_device_hwv(size_t device_index, int major, int minor, int minor2)
 {
   if (device_index >= (int)oc_number_of_devices()) {
-    PRINT("device_index %d to large\n", device_index);
+    PRINT("device_index %d to large\n", (int)device_index);
     return -1;
   }
 
@@ -261,10 +261,10 @@ oc_core_set_device_hwv(int device_index, int major, int minor, int minor2)
 }
 
 int
-oc_core_set_device_ia(int device_index, int ia)
+oc_core_set_device_ia(size_t device_index, int ia)
 {
   if (device_index >= (int)oc_number_of_devices()) {
-    PRINT("device_index %d to large\n", device_index);
+    PRINT("device_index %d to large\n", (int)device_index);
     return -1;
   }
   oc_device_info[device_index].ia = ia;
@@ -272,10 +272,10 @@ oc_core_set_device_ia(int device_index, int ia)
 }
 
 int
-oc_core_set_device_hwt(int device_index, const char *hardwaretype)
+oc_core_set_device_hwt(size_t device_index, const char *hardwaretype)
 {
   if (device_index >= (int)oc_number_of_devices()) {
-    PRINT("device_index %d to large\n", device_index);
+    PRINT("device_index %d to large\n", (int)device_index);
     return -1;
   }
 
@@ -287,10 +287,10 @@ oc_core_set_device_hwt(int device_index, const char *hardwaretype)
 }
 
 int
-oc_core_set_device_pm(int device_index, bool pm)
+oc_core_set_device_pm(size_t device_index, bool pm)
 {
   if (device_index >= (int)oc_number_of_devices()) {
-    PRINT("  device_index %d to large\n", device_index);
+    PRINT("  device_index %d to large\n", (int)device_index);
     return -1;
   }
 
@@ -300,10 +300,10 @@ oc_core_set_device_pm(int device_index, bool pm)
 }
 
 int
-oc_core_set_device_model(int device_index, const char *model)
+oc_core_set_device_model(size_t device_index, const char *model)
 {
   if (device_index >= (int)oc_number_of_devices()) {
-    PRINT("  device_index %d to large\n", device_index);
+    PRINT("  device_index %d to large\n", (int)device_index);
     return -1;
   }
   oc_free_string(&oc_device_info[device_index].model);
@@ -313,10 +313,10 @@ oc_core_set_device_model(int device_index, const char *model)
 }
 
 int
-oc_core_set_device_hostname(int device_index, const char *hostname)
+oc_core_set_device_hostname(size_t device_index, const char *hostname)
 {
   if (device_index >= (int)oc_number_of_devices()) {
-    PRINT("  device_index %d to large\n", device_index);
+    PRINT("  device_index %d to large\n", (int)device_index);
     return -1;
   }
   oc_free_string(&oc_device_info[device_index].hostname);
@@ -327,10 +327,10 @@ oc_core_set_device_hostname(int device_index, const char *hostname)
 }
 
 int
-oc_core_set_device_iid(int device_index, uint32_t iid)
+oc_core_set_device_iid(size_t device_index, uint32_t iid)
 {
   if (device_index >= (int)oc_number_of_devices()) {
-    PRINT("  device_index %d to large\n", device_index);
+    PRINT("  device_index %d to large\n", (int)device_index);
     return -1;
   }
   oc_device_info[device_index].iid = iid;
